# Improve text view performance