### Details
I've updated a few things in this PR.

#### Convert part of Storage.swift to ObjC
Yes, we've converted some code back to objc as it cuts some processing time. In case text has a lot of attributes bridging swift to objc becomes expensive.
My rough benchmark on a note with 40k characters, highlighting 2k words in it.
`noteEditor.highlightedRanges = ranges` line was taking:
Before: 0.1 sec
After: 0.014 sec

#### Wrap `checkTextInDocument` in begin / end editing
`checkTextInDocument` is used to render links and in some notes with a lot of links it takes a lot of time.
My rough benchmark on a note with 90k characters containing only links.
Measuring `processLinksInDocument`:
Before: 2.44 sec
After: 0.33 sec

### Test
Do a quick smoke test around note editor and working with links.

### Review
Only one developer is required to review these changes, but anyone can perform the review.

### Release
> These changes do not require release notes.