@eshurakov 2021-02-09
@mokagio could you make a new build please?

@pachlava could we bug you with this PR. It would be great to do some additional testing around opening big notes, searching and working with links.