@pachlava 2021-02-20
[testing] fix/feature verified on MBP 16" Intel | macOS 11.2 | Simplenote 2.8 beta3 (11220)